Added web api project

+++ InventorySystem.Manager/GuideManager.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using InventorySystem.Contract;
 using InventorySystem.Core;
 using InventorySystem.DataAccess.Implementation;
@@ -10,6 +11,7 @@ namespace InventorySystem.Manager
     public class GuideManager
     {
         private IGenericRepository<Product> _productRepository;
+        private IGenericRepository<ProductWork> _productWorkRepository;
         private IGenericRepository<Unit> _unitRepository;
         private IGenericRepository<Provider> _providerRepository;
         private IGenericRepository<PurchaseInvoice> _purchaseInvoiceRepository;
@@ -24,6 +26,7 @@ namespace InventorySystem.Manager
             _purchaseInvoiceRepository = RootContainer.Container.Resolve<IGenericRepository<PurchaseInvoice>>();
             _invoiceRepository = RootContainer.Container.Resolve<IGenericRepository<Invoice>>();
             _inventoryRepository = RootContainer.Container.Resolve<IGenericRepository<Inventory>>();
+            _productWorkRepository = RootContainer.Container.Resolve<IGenericRepository<ProductWork>>();
         }
 
         public void CreateProduct(Product item)
@@ -31,9 +34,17 @@ namespace InventorySystem.Manager
             _productRepository.Create(item);
         }
 
-        public IEnumerable<Product> GetProducts()
+        public IEnumerable<Product> GetProducts(bool showIsDelete = false)
         {
-            return _productRepository.GetWithInclude(x=>x.Unit);
+            if (showIsDelete)
+            {
+                return _productRepository.GetWithInclude(x => x.Unit);
+            }
+            else
+            {
+                return _productRepository.GetWithInclude(x => x.Unit)
+                    .Where(x => x.IsDelete == null || x.IsDelete == false);
+            }
         }
 
         public void CreateUnit(Unit item)
@@ -41,9 +52,16 @@ namespace InventorySystem.Manager
             _unitRepository.Create(item);
         }
 
-        public IEnumerable<Unit> GetUnits()
+        public IEnumerable<Unit> GetUnits(bool showIsDelete = false)
         {
-            return _unitRepository.Get();
+            if (showIsDelete)
+            {
+                return _unitRepository.Get();
+            }
+            else
+            {
+                return _unitRepository.Get().Where(x => x.IsDelete == null || x.IsDelete == false);
+            }
         }
 
         public void CreateProvider(Provider item)
@@ -51,9 +69,16 @@ namespace InventorySystem.Manager
             _providerRepository.Create(item);
         }
 
-        public IEnumerable<Provider> GetProviders()
+        public IEnumerable<Provider> GetProviders(bool showIsDelete = false)
         {
-            return _providerRepository.Get();
+            if (showIsDelete)
+            {
+                return _providerRepository.Get();
+            }
+            else
+            {
+                return _providerRepository.Get().Where(x => x.IsDelete == null || x.IsDelete == false);
+            }
         }
 
         public void UpdateProduct(Product item)
@@ -73,17 +98,45 @@ namespace InventorySystem.Manager
         
         public void DeleteProduct(Product item)
         {
-            _productRepository.Remove(item);
+            var relatives = _productWorkRepository.Get(x => x.ProductId == item.Id);
+            if (relatives.Any())
+            {
+                item.IsDelete = true;
+                item.Unit = null;
+                UpdateProduct(item);
+            }
+            else
+            {
+                _productRepository.Remove(item);
+            }
         }
 
         public void DeleteUnit(Unit item)
         {
-            _unitRepository.Remove(item);
+            var relatives = _productRepository.Get(x => x.UnitId == item.Id);
+            if (relatives.Any())
+            {
+                item.IsDelete = true;
+                UpdateUnit(item);
+            }
+            else
+            {
+                _unitRepository.Remove(item);
+            }
         }
 
         public void DeleteProvider(Provider item)
         {
-            _providerRepository.Remove(item);
+            var relatives = _purchaseInvoiceRepository.Get(x => x.ProviderId == item.Id);
+            if (relatives.Any())
+            {
+                item.IsDelete = true;
+                UpdateProvider(item);
+            }
+            else
+            {
+                _providerRepository.Remove(item);
+            }
         }
 
         public IEnumerable<PurchaseInvoice> GetPurchaseInvoices()
